🐛 Fix unit-test regression: use partial mock for wsAuth in kubectlProxy tests
Fixes #10765

The vi.mock for appendWsAuthToken was incomplete and caused imports of isDemoMode from the underlying demoMode module to fail with 'isDemoMode not defined'.

Changed to use async importActual to preserve other exports and only override appendWsAuthToken, ensuring dependencies are properly resolved.
